added poisson point process random input generator

+++ src/core/util/tools.h
@@ -22,7 +22,7 @@ inline float fSet(float* arr, int size, float val, float fraction=1.0) {
     if (fraction == 1.0) {
         for (int i = 0 ; i < size ; ++i) arr[i] = val;
     } else {
-        auto dist = std::uniform_real_distribution<double>(0.0, 1.0);
+        auto dist = std::uniform_real_distribution<float>(0.0, 1.0);
         for (int i = 0 ; i < size ; ++i)
             if (dist(generator) < fraction) arr[i] = val;
     }
@@ -30,16 +30,16 @@ inline float fSet(float* arr, int size, float val, float fraction=1.0) {
 
 /* Random float functions */
 inline float fRand() {
-    return std::uniform_real_distribution<double>(0.0, 1.0)(generator);
+    return std::uniform_real_distribution<float>(0.0, 1.0)(generator);
 }
 inline float fRand(float fMax) {
-    return std::uniform_real_distribution<double>(0.0, fMax)(generator);
+    return std::uniform_real_distribution<float>(0.0, fMax)(generator);
 }
 inline float fRand(float fMin, float fMax) {
-    return std::uniform_real_distribution<double>(fMin, fMax)(generator);
+    return std::uniform_real_distribution<float>(fMin, fMax)(generator);
 }
 inline void fRand(float* arr, int size, float fMin, float fMax, float fraction=1.0) {
-    auto dist = std::uniform_real_distribution<double>(fMin, fMax);
+    auto dist = std::uniform_real_distribution<float>(fMin, fMax);
     if (fraction == 1.0)
         for (int i = 0 ; i < size ; ++i) arr[i] = dist(generator);
     else
@@ -63,7 +63,7 @@ inline void iRand(int* arr, int size, int iMin, int iMax, float fraction=1.0) {
     if (fraction == 1.0)
         for (int i = 0 ; i < size ; ++i) arr[i] = dist(generator);
     else {
-        auto f_dist = std::uniform_real_distribution<double>(0.0, 1.0);
+        auto f_dist = std::uniform_real_distribution<float>(0.0, 1.0);
         for (int i = 0 ; i < size ; ++i)
             if (f_dist(generator) < fraction)
                 arr[i] = dist(generator);
